Cascade
To fix the paragraphs we needed to move small-para so it would execute last according to the rule order
One of the buttons needed to be selected, by adding the confirm-button ID selector, we specified its selection.
The text elements required us to move the child div last and recall its ancestor by using a descendant combinator.